+++ src/common/components/ArmoryBadge/index.js
@@ -4,7 +4,7 @@ import styles from './styles.less';
 
 const ArmoryBadge = () => (
   <div className={styles.root}>
-    <a href="https://gw2armory.com">gw2armory.com &#9829;</a>
+    <a href="https://gw2armory.com">powered by gw2armory.com &#9829;</a>
   </div>
 );